Take /etc/fstab into account when mounting pseudo filesystems.
First try remounting the pseudofs with fstab options, if this fails we
assume it is not mounted and try mounting it with options from fstab, if
this again fails, the pseudofs cannot be in fstab so we try mounting it
with default options. All error messages except for the last try are
suppressed.